Arduino analog pins 4 and 5 mode switcher redesigned.
The N/P MOSFET pairs (Q15 and Q16) would not have functioned correctly in a reverse voltage/current condition.
They have been replaced with 2:1 mux/demux ICs (IC4 and IC5) which are a much better fit for the configuration.
Q31 and Q32 have been renamed to Q15 and Q16 respectively to remove the gap in Q__ part names.