Updated gyro U14 to new model, the L3GD20.
C27 and R40 are noted in the schematic as DNP unless using an L3G4200D (they are pin and register compatible). They have also been removed from the BOM as the use of a L3G4200D will not be a standard configuration.
Added WHO_AM_I and slave address definitions to L3G4200D.h for both gyros.